% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emit_enum.R
\name{emit_enums_file}
\alias{emit_enums_file}
\title{The source of \code{R/enums-generated.R}}
\usage{
emit_enums_file(enums)
}
\arguments{
\item{enums}{List of \code{devforge_enum} objects.}
}
\value{
Character vector of R source lines, without the generated header.
}
\description{
The source of \code{R/enums-generated.R}
}
\keyword{internal}
