% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{indent}
\alias{indent}
\title{Indent a block of source code}
\usage{
indent(lines, n = 2L)
}
\arguments{
\item{lines}{Character vector. The lines to indent.}

\item{n}{Integer. Number of spaces. Defaults to \code{2L}.}
}
\value{
Character vector with each non-empty line indented.
}
\description{
Indent a block of source code
}
\keyword{internal}
