% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_lines}
\alias{print_lines}
\title{Print the format() lines of an object}
\usage{
print_lines(x, ...)
}
\arguments{
\item{x}{Any object with a \code{\link[=format]{format()}} method returning lines.}

\item{...}{Passed on to \code{\link[=format]{format()}}.}
}
\value{
\code{x}, invisibly.
}
\description{
Print the format() lines of an object
}
\keyword{internal}
