% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{run_air}
\alias{run_air}
\title{Run air over a set of files}
\usage{
run_air(paths)
}
\arguments{
\item{paths}{Character vector of file paths.}
}
\value{
\code{TRUE} if air ran, \code{FALSE} if it is not on the path.
}
\description{
Run air over a set of files
}
\keyword{internal}
