% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{section_header}
\alias{section_header}
\title{A section divider padded to the 80 character line width}
\usage{
section_header(title)
}
\arguments{
\item{title}{String. The section title.}
}
\value{
String. The divider line.
}
\description{
A section divider padded to the 80 character line width
}
\keyword{internal}
