% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{spec_field_names}
\alias{spec_field_names}
\title{The names of a spec's returned list, in order}
\usage{
spec_field_names(spec)
}
\arguments{
\item{spec}{A \code{devforge_spec}.}
}
\value{
Character vector of field names.
}
\description{
The names of a spec's returned list, in order
}
\keyword{internal}
