% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{variant_class}
\alias{variant_class}
\title{The class of one enum variant}
\usage{
variant_class(enum, variant)
}
\arguments{
\item{enum}{A \code{devforge_enum}.}

\item{variant}{String. The variant name.}
}
\value{
String. \code{"<Enum>_<Variant>"}.
}
\description{
The class of one enum variant
}
\keyword{internal}
