% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_devforge}
\alias{print_devforge}
\alias{print.devforge_field}
\alias{print.devforge_variant}
\alias{print.devforge_enum}
\alias{print.devforge_spec}
\title{Print a devforge object}
\usage{
\method{print}{devforge_field}(x, ...)

\method{print}{devforge_variant}(x, ...)

\method{print}{devforge_enum}(x, ...)

\method{print}{devforge_spec}(x, ...)
}
\arguments{
\item{x}{A \code{devforge_field}, \code{devforge_variant}, \code{devforge_enum} or
\code{devforge_spec}.}

\item{...}{Passed on to \code{\link[=format]{format()}}.}
}
\value{
\code{x}, invisibly.
}
\description{
Prints the \code{\link[=format]{format()}} lines of a field, variant, enum or spec.
}
